Project Structure Update
