+++ lib/services/auth_service.dart
@@ -1,0 +1,30 @@
+import 'package:google_sign_in/google_sign_in.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+
+class AuthenticationService {
+  static Future<String> signInWithGoogle() async {
+    final GoogleSignInAccount googleSignInAccount =
+        await GoogleSignIn().signIn();
+    final GoogleSignInAuthentication googleSignInAuthentication =
+        await googleSignInAccount.authentication;
+
+    final AuthCredential credential = GoogleAuthProvider.credential(
+      accessToken: googleSignInAuthentication.accessToken,
+      idToken: googleSignInAuthentication.idToken,
+    );
+
+    final UserCredential authResult =
+        await FirebaseAuth.instance.signInWithCredential(credential);
+    final User user = authResult.user;
+    assert(!user.isAnonymous);
+
+    final User currentUser = FirebaseAuth.instance.currentUser;
+    assert(user.uid == currentUser.uid);
+
+    return '$user';
+  }
+
+  static Future<void> signOutGoogle() async {
+    await GoogleSignIn().signOut();
+  }
+}
