--- lib/services/google_auth.dart
@@ -1,32 +1,0 @@
-import 'package:firebase_core/firebase_core.dart';
-import 'package:google_sign_in/google_sign_in.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-
-final FirebaseAuth _auth = FirebaseAuth.instance;
-final GoogleSignIn googleSignIn = GoogleSignIn();
-
-
-Future<String> signInWithGoogle() async {
-  await Firebase.initializeApp();
-
-  final GoogleSignInAccount googleSignInAccount = await googleSignIn.signIn();
-  final GoogleSignInAuthentication googleSignInAuthentication = await googleSignInAccount.authentication;
-
-  final AuthCredential credential = GoogleAuthProvider.credential(
-    accessToken: googleSignInAuthentication.accessToken,
-    idToken: googleSignInAuthentication.idToken,
-  );
-
-  final UserCredential authResult = await _auth.signInWithCredential(credential);
-  final User user = authResult.user;
-  assert(!user.isAnonymous);
-
-  final User currentUser = _auth.currentUser;
-  assert(user.uid == currentUser.uid);
-
-  return '$user';
-}
-
-Future<void> signOutGoogle() async {
-  await googleSignIn.signOut();
-}
