+++ lib/views/profile/settings/faqs_screen.dart
