+++ lib/views/profile/settings/manage_address.dart
@@ -3,7 +3,7 @@ import 'package:geocoder/geocoder.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:retro_shopping/helpers/constants.dart';
 import 'package:retro_shopping/model/address.dart';
-import 'package:retro_shopping/views/address_screen.dart';
+import 'package:retro_shopping/views/profile/settings/address_screen.dart';
 import 'package:retro_shopping/widgets/address_text_field.dart';
 import 'package:retro_shopping/widgets/retro_button.dart';
