+++ pubspec.yaml
@@ -24,12 +24,8 @@ dependencies:
   flutter:
     sdk: flutter
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  adobe_xd: ^1.1.0+1
-  cupertino_icons: ^0.1.3
   fluttertoast: ^7.1.8
-  firebase_core:
+  firebase_core: 0.7.0
   google_fonts: ^1.1.0
   razorpay_flutter: 1.1.0
   google_mobile_ads: ^0.11.0+3
@@ -38,8 +34,8 @@ dependencies:
   flutter_svg: ^0.20.0-nullsafety.3
   provider:
   advance_pdf_viewer:
-  firebase_auth:
-  google_sign_in:
+  firebase_auth: 0.20.1
+  google_sign_in: 4.5.9
 
 dev_dependencies:
   flutter_test:
@@ -59,7 +55,6 @@ flutter:
   assets:
     - assets/
     - assets/items/
-    - images/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
